project_name/requirements/base.txt
#Core
Django

#WSGI HTTP Server
gunicorn
pytz

#Configuration
django-environ
pillow

#Database
psycopg2


#Redis
django-redis
redis

#Celery
celery==3.1.24


#Social
django-allauth



django-storages-redux

#API
django-cors-headers
djangorestframework
djangorestframework-jsonapi
requests
